refactore to instant.library.ts
