--- src/app/route/forum/forum.define.ts
@@ -1,44 +1,0 @@
-
-export const CLEAN = {
-	gift1: /Gift #\d+,/,
-	gift2: /, Gift #\d+/,
-	gift3: /Gift #\d+/,
-	and: /and #\d+/,
-	week: /Bonus: Week Level reached/,
-	week2: /Bonus: week level reached/,
-	spaces: /  +/g,
-	newline: /\n/g,
-	newlines: /\n\s*\n/g,
-	comma1: /^,/,
-	comma2: /,$/,
-	colon1: /^:/,
-	colon2: /:$/,
-}
-export const COMMENTS = [
-	'Great',
-	'Hello',
-	'Happy',
-	'Jorge',
-	'love',
-	'good',
-	'relax',
-	"I'm",
-	'I am',
-	'Something',
-	'Instructor',
-	'routine',
-	'Whateva',
-	'holiday',
-	'clever',
-	'xxx',
-	' 2017',
-	'Lucie',
-	"'",
-	'"',
-	',',
-	'.',
-	'!',
-	'?',
-	':',
-	';'
-];
