+++ src/app/service/admin/admin.service.ts
@@ -7,7 +7,7 @@ import { MemberService } from '@service/member/member.service';
 import { FIELD, STORE } from '@dbase/data/data.define';
 import { IStoreMeta, IGift } from '@dbase/data/data.schema';
 
-import { getStamp, TDate, getDate } from '@lib/date.library';
+import { getStamp, TDate, getDate } from '@lib/instant.library';
 import { TString } from '@lib/type.library';
 import { asArray } from '@lib/array.library';
 import { dbg } from '@lib/logger.library';
